requirements.txt
behave~=1.2.6
selenium~=4.25.0
jsonschema~=4.23.0
schema~=0.7.7
webdriver-manager~=4.0.2
requests~=2.32.3
factory_boy~=3.3.1
Faker~=30.3.0